requirements.txt
flask>=3.0
gunicorn>=22.0
pandas>=2.2
openpyxl>=3.1
python-pptx>=1.0
matplotlib>=3.9
google-cloud-storage>=2.18
